add lastmodifiedAt on stock detail page
