+++ src/app/(main)/report/sector-detail/_components/sector-insights.tsx
@@ -1,25 +1,11 @@
 import React from "react";
 import { SectorInsightsItem } from "./sector-insights-item";
 import { StockDividendYieldResponse, UpcomingDividendResponse } from "../../../../../api/generated/endpoint.schemas";
+import { formatDate } from "../../../../../utils/date";
 
 interface SectorInsightsProps {
   comingDividendStocks?: UpcomingDividendResponse;
   biggestDividendYieldStocks?: StockDividendYieldResponse;
-}
-
-function formatDate(dateString: string) {
-  try {
-    const date = new Date(dateString);
-    const year = date.getFullYear();
-    const month = (date.getMonth() + 1).toString().padStart(2, "0");
-    const day = date.getDate().toString().padStart(2, "0");
-    const hours = date.getHours().toString().padStart(2, "0");
-    const minutes = date.getMinutes().toString().padStart(2, "0");
-
-    return `${year}/${month}/${day} ${hours}:${minutes}`;
-  } catch {
-    return undefined;
-  }
 }
 
 export const SectorInsights = React.memo(
